make notyfyCustomButton
# Conflicts:
#	.idea/deploymentTargetDropDown.xml
#	app/src/main/java/com/example/ManabaApp/AddTaskCustomDialog.java
#	app/src/main/java/com/example/ManabaApp/ChangeableClassDialog.java
#	app/src/main/java/com/example/ManabaApp/ClassData.java
#	app/src/main/java/com/example/ManabaApp/ClassDataManager.java
#	app/src/main/java/com/example/ManabaApp/ClassGridAdapter.java
#	app/src/main/java/com/example/ManabaApp/DataManager.java
#	app/src/main/java/com/example/ManabaApp/LogInManager.java
#	app/src/main/java/com/example/ManabaApp/LoginActivity.java
#	app/src/main/java/com/example/ManabaApp/LoginDialog.java
#	app/src/main/java/com/example/ManabaApp/MainActivity.java
#	app/src/main/java/com/example/ManabaApp/ManabaScraper.java
#	app/src/main/java/com/example/ManabaApp/NotificationCustomAdapter.java
#	app/src/main/java/com/example/ManabaApp/NotificationReceiver2.java
#	app/src/main/java/com/example/ManabaApp/RegisterClassDialog.java
#	app/src/main/java/com/example/ManabaApp/TaskDataManager.java
#	app/src/main/java/com/example/ManabaApp/UnChangeableClassDialog.java

+++ app/src/main/java/com/example/ManabaApp/MainActivity.java
@@ -156,9 +156,13 @@ public class MainActivity extends AppCompatActivity implements ClassUpdateListen
                         LogOffButton.setOnClickListener(new View.OnClickListener() {//ボタンが押されたら
                             @Override
                             public void onClick(View v) {//ボタンが押されたら
-                                cookieManager.removeAllCookie();
-                                db.execSQL("DELETE FROM " + "TaskData");
-                                db.execSQL("DELETE FROM " + "ClassData");
+                                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                                    cookieBag.entrySet().forEach(entry -> {
+                                    });
+                                    cookieManager.removeAllCookie();
+                                    db.execSQL("DELETE FROM " + "TaskData");
+                                    db.execSQL("DELETE FROM " + "ClassData");
+                                }
                             }
                         });
